only show filter when there's results
