+++ catalogue/webapp/components/WorkCompactCard/WorkCompactCard.js
