+++ catalogue/webapp/services/catalogue/works.js
@@ -1,5 +1,11 @@
 // @flow
 import fetch from 'isomorphic-unfetch';
+import {
+  type CatalogueResultsList,
+  type CatalogueApiError,
+  type Work,
+  type CatalogueApiRedirect,
+} from '@weco/common/model/catalogue';
 
 type GetWorksProps = {|
   query: string,
@@ -10,34 +16,6 @@ type GetWorksProps = {|
 type GetWorkProps = {|
   id: string,
 |};
-
-export type Work = {
-  type: 'Work',
-  ...Object,
-};
-
-export type CatalogueApiError = {|
-  errorType: string,
-  httpStatus: number,
-  label: string,
-  description: string,
-  type: 'Error',
-|};
-
-export type CatalogueResultsList = {
-  type: 'ResultList',
-  totalResults: number,
-  results: Work[],
-  pageSize: number,
-  prevPage: ?string,
-  nextPage: ?string,
-};
-
-export type CatalogueApiRedirect = {
-  type: 'Redirect',
-  status: number,
-  redirectToId: string,
-};
 
 const rootUri = 'https://api.wellcomecollection.org/catalogue';
 const includes = [
